Mc1.12.x sync

+++ src/main/java/moze_intel/projecte/gameObjs/gui/GUICollectorMK1.java
@@ -34,9 +34,9 @@ public class GUICollectorMK1 extends GuiContainer
 	@Override
 	protected void drawGuiContainerForegroundLayer(int var1, int var2)
 	{
-		this.fontRenderer.drawString(Integer.toString(container.emc), 60, 32, 4210752);
+		this.fontRenderer.drawString(Long.toString(container.emc), 60, 32, 4210752);
 		
-		double kleinCharge = container.kleinEmc;
+		long kleinCharge = container.kleinEmc;
 
 		if (kleinCharge > 0)
 			this.fontRenderer.drawString(Constants.EMC_FORMATTER.format(kleinCharge), 60, 44, 4210752);
